workshop/week3/requirements.txt
python-dotenv==1.1.0
pypdf==5.4.0
tavily-python==0.5.4

langchain==0.3.21
langchain-core==0.3.49
langchain-community== 0.3.20
langchain-openai== 0.3.11
langchain-chroma==0.2.2
langchain-text-splitters==0.3.7
langchainhub== 0.1.21
langgraph==0.3.21

langsmith==0.3.19

chainlit==2.4.302